% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_allometry_matrix.R
\name{add_missing_bwg_names}
\alias{add_missing_bwg_names}
\title{Add missing names}
\usage{
add_missing_bwg_names(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\description{
Add a placeholder for species missing a BWG name. Makes it possible to
group by bwg_name even when one wasn't given.
}
\keyword{internal}
