% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanup_internal.R
\name{calculate_biomass}
\alias{calculate_biomass}
\title{Calculate biomass}
\usage{
calculate_biomass(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\description{
Calculate biomass. When multiple mass measurements per species-length are
included, the weighted average is taken.
}
\keyword{internal}
