% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equation_bank_internal.R
\name{clean_allometry_matrix}
\alias{clean_allometry_matrix}
\title{Clean allometry matrix}
\usage{
clean_allometry_matrix(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\description{
Remove rows with missing mass or length
}
\keyword{internal}
