% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biomass_table_internal.R
\name{clean_lengths}
\alias{clean_lengths}
\title{Clean lengths}
\usage{
clean_lengths(biomass_table)
}
\arguments{
\item{biomass_table}{The biomass table}
}
\description{
Takes strings that should be lengths, cleans them up, and converts them
accordingly.
}
\keyword{internal}
