% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biomass_table_internal.R
\name{correct_size_categories}
\alias{correct_size_categories}
\title{Correct size categories}
\usage{
correct_size_categories(biomass_data)
}
\arguments{
\item{biomass_data}{The biomass table}
}
\description{
Correct NA size categories - generally replacing NA or "regular" with
"average".
}
\keyword{internal}
