% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomass_table.R
\name{correct_stages}
\alias{correct_stages}
\title{Correct stages}
\usage{
correct_stages(biomass_data)
}
\arguments{
\item{biomass_data}{The biomass table}
}
\description{
Correct stage errors. Stage "default" is converted to "larva" to match the
allometry matrices. Species with stage "larva" and size category NA are
given size category "average". Some rows have "stages" that are actually a
size category. For example, the stage is listed as "small" instead of
"larva".  In these rows, the erroneous stage is moved into the
"size_catgory" column, and "stage" is assumed to be "larva".
}
\keyword{internal}
