% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_allometry_matrix.R
\name{create_allometry_matrix}
\alias{create_allometry_matrix}
\title{Create the allometry matrix}
\usage{
create_allometry_matrix()
}
\value{
The allometry matrix as a data frame.
}
\description{
Creates the allometry matrix from the allometry data set:
  \url{https://github.com/SrivastavaLab/allometrydata/}
}
