% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomass_table.R
\name{create_biomass_table}
\alias{create_biomass_table}
\title{Create the cleaned up biomass table}
\usage{
create_biomass_table(species_ids = NULL)
}
\arguments{
\item{species_ids}{A table of species_ids and corresponding taxonomy info
created by using the function "get_species_ids()" in this package. Default
value of "NULL".}
}
\description{
Create the cleaned up biomass table by importing the existing biomass
data and removing existing biomass information.
}
