% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_category_lookup.R
\name{create_category_lookup}
\alias{create_category_lookup}
\title{Create category lookup table}
\usage{
create_category_lookup()
}
\arguments{
\item{category_lookup}{The category lookup table.}
}
\value{
The category lookup table as a data frame.
}
\description{
Create the category lookup table. This matches a size category to a length
estimate so that biomasses can be calculated on categorical sizes.
}
