% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equation_bank.R
\name{create_equation_bank}
\alias{create_equation_bank}
\title{Create equation bank}
\usage{
create_equation_bank(allometry_matrix = NULL)
}
\arguments{
\item{allometry_matrix}{The allometry matrix as created by this package
 using the "create_allometry_matrix()" function. Default value is NULL. If
 left as NULL, the function will start by creating the allometry matrix.

 @return A data frame containing the equation bank}
}
\description{
Create the allometric equation bank
}
