% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{filter_by_biomass_type}
\alias{filter_by_biomass_type}
\title{Filter by biomass type}
\usage{
filter_by_biomass_type(biomass_data)
}
\arguments{
\item{biomass_data}{The cleaned biomass table}
}
\description{
Filter based on whether the measurement is for dry or wet biomass.
}
\keyword{internal}
