% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanup_internal.R
\name{filter_type}
\alias{filter_type}
\title{Filter by data type}
\usage{
filter_type(allometry_data, measurement = TRUE)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}

\item{measurement}{logical indicating whether the measurements should be filtered into
the final data set (type = field_measurement). Otherwise, the category
definitions will be filtered into the final data set
(type = category_definition). Defaults to \emph{measurement = TRUE}}
}
\value{
the filtered data set
}
\description{
Filter by data type.
}
\keyword{internal}
