% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equation_bank_internal.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(x)
}
\arguments{
\item{x}{Data on which to perform the model fit}
}
\description{
The model fit function.
}
\keyword{internal}
