% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdata_create_internal.R
\name{fwdata_create}
\alias{fwdata_create}
\title{Create a new fwdata release}
\usage{
fwdata_create(data_dir = "_data", version)
}
\arguments{
\item{data_dir}{A local data directory in which to save the fwdata including
the biomass data frame}

\item{version}{Optional data version number}
}
\description{
Create a new fwdata release using the most up to date version
(\url{https://github.com/SrivastavaLab/cesabfunctionalwebsdata/})
}
\keyword{internal}
