% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_biomass.R
\name{gen_biomass}
\alias{gen_biomass}
\title{Generate biomass estimates}
\usage{
gen_biomass(data_dir = "_data", version = NULL)
}
\arguments{
\item{data_dir}{A local data directory in which to save the biomass data}

\item{version}{Optional data version number}
}
\description{
Generate biomass estimates
(\url{https://github.com/SrivastavaLab/bwgbiomass/}) from allometry data
 (\url{https://github.com/SrivastavaLab/allometrydata/})
}
