% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{generate_biomass}
\alias{generate_biomass}
\title{Generate biomass}
\usage{
generate_biomass(biomass_data, meas_ids, prov, prov_species,
  allometry_data, equation_bank)
}
\arguments{
\item{biomass_data}{The biomass data, with all current biomass estimates}

\item{meas_ids}{IDs of measurements that already have biomass}

\item{prov}{Provenance of the biomass measurement}

\item{prov_species}{Provenance determined for the species in question, or
the closest related specices}

\item{allometry_data}{The allometry matrix}

\item{equation_bank}{The equation bank}
}
\description{
Generate biomass calculation for given measurement. Provenance information
is added at this time.
}
\keyword{internal}
