% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{generate_biomass_fun}
\alias{generate_biomass_fun}
\title{Generate biomass function}
\usage{
generate_biomass_fun(BIOMASS, ALLOMETRY, EQUATIONS)
}
\arguments{
\item{BIOMASS}{The biomass data, with all current biomass estimates}

\item{ALLOMETRY}{The allometry matrix}

\item{EQUATIONS}{The equation bank}

\item{prov_species}{Provenance determined for the species in question, or
the closest related specices}

\item{prov_meas}{Provenance measurement (length or category)}

\item{prov_type}{Provenance type (raw or interpolate)}
}
\value{
A function for generating biomass calculations
}
\description{
Generate a function for making the biomass calculation for a given
measurement. Provenance information is added at this time.
}
\keyword{internal}
