% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biomass_table_internal.R
\name{generate_biomass_table}
\alias{generate_biomass_table}
\title{Generate the biomass table}
\usage{
generate_biomass_table(species_info)
}
\arguments{
\item{species_info}{A table of species_ids and corresponding taxonomy info}
}
\description{
Obtain the original biomass table from the saved package data. Clean
the table for further use. The original table is called "biomass_original"
and is stored as system data (sysdata).
}
\keyword{internal}
