% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{generate_median_biomass}
\alias{generate_median_biomass}
\title{Generate median biomass}
\usage{
generate_median_biomass(biomass_data, allometry_species, equation_species)
}
\arguments{
\item{biomass_data}{The cleaned biomass table}
}
\description{
Generate median biomass.
}
\keyword{internal}
