% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_relatives_internal.R
\name{get_closest_relatives}
\alias{get_closest_relatives}
\title{Get closest relatives}
\usage{
get_closest_relatives(biomass_data, allometry_data, species_data)
}
\arguments{
\item{biomass_data}{The biomass table}

\item{allometry_data}{The cleaned up allometry data}

\item{species_data}{Species ID data with taxonomy attached}
}
\description{
Create a table of the closest taxonomic relatives to any given species.
}
\keyword{internal}
