% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{get_median_species}
\alias{get_median_species}
\title{Get median species}
\usage{
get_median_species(shared_species)
}
\arguments{
\item{shared_species}{The vector of shared species}
}
\description{
Get the identity of species with median biomass
}
\keyword{internal}
