% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_ids_internal.R
\name{get_species_ids}
\alias{get_species_ids}
\title{Get Species IDs}
\usage{
get_species_ids()
}
\description{
Get species taxonomy information from BWGdb to go with the species ID.
}
\keyword{internal}
