% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{interpolate_biomass}
\alias{interpolate_biomass}
\title{Interpolate biomass}
\usage{
interpolate_biomass(biomass_data, type)
}
\arguments{
\item{biomass_data}{The biomass table}

\item{type}{Interpolate using "length" or "category"}
}
\description{
Get the identity of species with median biomass
}
\keyword{internal}
