% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biomass_table.R
\name{obtain_biomass_table}
\alias{obtain_biomass_table}
\title{Obtain original biomass table}
\usage{
obtain_biomass_table(species_info)
}
\arguments{
\item{species_info}{A table of species_ids and corresponding taxonomy info}
}
\description{
Obtain the original biomass table from the saved package data. Clean
the table for further use. The original table is called "biomass_original"
and is stored as system data (sysdata).
}
\keyword{internal}
