% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equation_bank_internal.R
\name{obtain_summary_data}
\alias{obtain_summary_data}
\title{Obtain summary data}
\usage{
obtain_summary_data(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\description{
Create the final equation bank by getting summary data.
}
\keyword{internal}
