% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_biomass_table_internal.R
\name{populate_biomass_table}
\alias{populate_biomass_table}
\title{Populate the biomass table}
\usage{
populate_biomass_table(allometry_data, equation_bank, biomass_data,
  species_data, category_data, closest_relatives)
}
\arguments{
\item{allometry_data}{The allometry matrix}

\item{equation_bank}{Allometric equation bank}

\item{biomass_data}{The biomass table}

\item{species_data}{Species IDs with taxonomy}

\item{category_data}{The category lookup table}

\item{closest_relatives}{The table of closest relatives}
}
\description{
Populate the biomass table with updated biomass estimates and provenance
information.
}
\keyword{internal}
