% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_allometry_matrix.R
\name{remove_file_col}
\alias{remove_file_col}
\title{Remove file column}
\usage{
remove_file_col(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\value{
Data with file column removed.
}
\description{
Remove the "file" column.
}
\keyword{internal}
