% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_allometry_matrix.R
\name{remove_missing_biomass}
\alias{remove_missing_biomass}
\title{Remove rows with no biomass}
\usage{
remove_missing_biomass(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\value{
Data with missing biomasses filtered out.
}
\description{
Remove rows with no biomass.
}
\keyword{internal}
