% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equation_bank.R
\name{run_linear_models}
\alias{run_linear_models}
\title{Run linear models}
\usage{
run_linear_models(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\description{
Run the linear models to create equations for calculating biomass.
}
\keyword{internal}
