% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Sanity check biomass measurements}
\usage{
sanity_check(biomass_data)
}
\arguments{
\item{biomass_data}{A table of biomass estimates. Must include the
measurement_id.}
}
\value{
A data.frame comparing new biomass estimates with the original
}
\description{
Compares the calculated biomass measurements with those found in the
original biomass table.
}
