% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_category_lookup_internal.R
\name{select_category_cols}
\alias{select_category_cols}
\title{Select category cols}
\usage{
select_category_cols(allometry_data)
}
\arguments{
\item{category_lookup}{The category lookup table.}
}
\description{
Select the relevant columns for the category lookup table. length_mm is
renamed to length_est_mm as the lengths are estimates for category
definitions.
}
\keyword{internal}
