% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_category_lookup_internal.R
\name{standardize_size_categories}
\alias{standardize_size_categories}
\title{Standardize size categories.}
\usage{
standardize_size_categories(category_lookup)
}
\arguments{
\item{category_lookup}{The category lookup table.}
}
\description{
Standardize the size categories in the category lookup table.
}
\keyword{internal}
