% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_allometry_matrix.R
\name{standardize_stages}
\alias{standardize_stages}
\title{Standardize developmental stage}
\usage{
standardize_stages(allometry_data)
}
\arguments{
\item{allometry_data}{The allometry data set loaded from:
\url{https://github.com/SrivastavaLab/allometrydata/}}
}
\value{
The data with stages standardized.
}
\description{
Standardize developmental stages. Possible stages are: "larva", "pupa",
"adult" or "NA".
}
\keyword{internal}
