% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biomass_table_internal.R
\name{all_ostracods_adults}
\alias{all_ostracods_adults}
\title{All ostracods adults}
\usage{
all_ostracods_adults(biomass_data)
}
\arguments{
\item{biomass_data}{The biomass table}
}
\description{
Set stage for all Ostracoda to "adult"
}
\keyword{internal}
