% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_biomass.R
\name{gen_biomass}
\alias{gen_biomass}
\title{Generate biomass estimates}
\usage{
gen_biomass()
}
\description{
Generate biomass estimates
(\url{https://github.com/SrivastavaLab/bwgbiomass/}) from allometry data
 (\url{https://github.com/SrivastavaLab/allometrydata/})
}
